refactor(v3/organizations): move serializers to serializers/column_mappings.py
